"muestra solo las que entran a su almacen"

+++ apiLogisticaAlmacen/routes/api/users_warehouses.js
@@ -2,7 +2,7 @@ const router = require('express').Router();
 
 const { checkRole } = require('../../helpers/middlewares/user.middleware');
 const { serverError } = require('../../helpers/validators');
-const { create, getById, getAll, deleteById, deleteWarehousesByUserId } = require('../../models/users_warehouses.model');
+const { create, getById, getAll, deleteById, deleteWarehousesByUserId,getbyuserId } = require('../../models/users_warehouses.model');
 
 
 router.get('/', async (req, res) => {
@@ -61,5 +61,15 @@ router.delete('/:userWarehouseId',
     serverError(res, error.message);
   }
 })
+
+router.get ('/:userId', async (req, res) => {
+  try {
+    const { userId } = req.params
+    const usersWarehouses = await getbyuserId(userId);
+    res.json(usersWarehouses);
+  } catch (error) {
+    serverError(res, error.message);
+  }
+});
 
 module.exports = router;
